Abstract buttons into their own components
